Removed unnecessary type cast

+++ crates/blockchain/blockchain.rs
@@ -317,7 +317,7 @@ impl Blockchain {
         let mut throughput = 0.0;
         if elapsed_seconds != 0.0 && total_gas_used != 0 {
             let as_gigas = (total_gas_used as f64).div(10_f64.powf(9_f64));
-            throughput = (as_gigas) / (elapsed_seconds as f64);
+            throughput = (as_gigas) / elapsed_seconds;
         }
 
         metrics!(
